+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `IrVariable.toIrGetValue(): IrGetValue`
 - `FirBasedSymbol<*>.valueParameters`
 - `nothingFirAnonymousFunctionExpression`
 - multiple safe FIR functions based on getting nullable types